Move connection functions to separate file
